expanding afsm
